Terraform script for modular server setup with docker and ansible
The script uses a server provider to build any number of servers,
of any type, name and in any location.
The server module defaults to:
* cx22 type
* nbg1 datacenter
* ubuntu-24.04 image

After creating the server, the script uses two provisioners, one
for docker the other for ansible.
It creates an ansible non root user with priviledges to run
docker and python